+++ py/conftest.py
@@ -44,6 +44,7 @@ drivers = (
     'Remote',
     'Safari',
     'WebKitGTK',
+    'ChromiumEdge',
 )
 
 
@@ -117,6 +118,9 @@ def driver(request):
             options = get_options('Firefox', request.config)
         if driver_class == 'WebKitGTK':
             options = get_options(driver_class, request.config)
+        if driver_class == 'ChromiumEdge':
+            options = get_options('Edge', request.config)
+            kwargs.update({'is_legacy': False})
         if driver_path is not None:
             kwargs['executable_path'] = driver_path
         if options is not None:
@@ -131,6 +135,10 @@ def get_options(driver_class, config):
     browser_path = config.option.binary
     browser_args = config.option.args
     options = None
+
+    if driver_class == 'Edge':
+        return getattr(webdriver, '{}Options'.format(driver_class))(False)
+
     if browser_path or browser_args:
         options = getattr(webdriver, '{}Options'.format(driver_class))()
         if driver_class == 'WebKitGTK':
