+++ py/selenium/webdriver/chromium/remote_connection.py
@@ -18,7 +18,7 @@
 from selenium.webdriver.remote.remote_connection import RemoteConnection
 
 
-class ChromeRemoteConnection(RemoteConnection):
+class ChromiumRemoteConnection(RemoteConnection):
 
     def __init__(self, remote_server_addr, keep_alive=True):
         RemoteConnection.__init__(self, remote_server_addr, keep_alive)
