+++ py/selenium/webdriver/edge/service.py
@@ -15,43 +15,40 @@
 # specific language governing permissions and limitations
 # under the License.
 
-from selenium.webdriver.common import service
+from selenium.webdriver.chromium import service
 
 
-class Service(service.Service):
+class Service(service.ChromiumService):
 
-    def __init__(self, executable_path, port=0, verbose=False, log_path=None):
+    def __init__(self, executable_path, port=0, verbose=False, log_path=None, is_legacy=True, service_args=None, env=None):
         """
         Creates a new instance of the EdgeDriver service.
-
         EdgeDriver provides an interface for Microsoft WebDriver to use
         with Microsoft Edge.
-
-        :param executable_path: Path to the Microsoft WebDriver binary.
-        :param port: Run the remote service on a specified port.
-            Defaults to 0, which binds to a random open port of the
-            system's choosing.
-        :verbose: Whether to make the webdriver more verbose (passes the
-            --verbose option to the binary). Defaults to False.
-        :param log_path: Optional path for the webdriver binary to log to.
-            Defaults to None which disables logging.
-
+        :Args:
+         - executable_path : Path to the Microsoft WebDriver binary.
+         - port : Run the remote service on a specified port. Defaults to 0, which binds to a random open port
+           of the system's choosing.
+         - verbose : Whether to make the webdriver more verbose (passes the --verbose option to the binary).
+           Defaults to False. Should be only used for legacy mode.
+         - log_path : Optional path for the webdriver binary to log to. Defaults to None which disables logging.
+         - is_legacy : Whether to use MicrosoftWebDriver.exe (legacy) or MSEdgeDriver.exe (chromium-based). Defaults to True.
+         - service_args : List of args to pass to the WebDriver service.
         """
+        self.service_args = service_args or []
 
-        self.service_args = []
-        if verbose:
-            self.service_args.append("--verbose")
+        if is_legacy:
+            if verbose:
+                self.service_args.append("--verbose")
 
-        params = {
-            "executable": executable_path,
-            "port": port,
-            "start_error_message": "Please download from https://go.microsoft.com/fwlink/?LinkId=619687"
-        }
+            if log_path:
+                params["log_file"] = open(log_path, "a+")
 
-        if log_path:
-            params["log_file"] = open(log_path, "a+")
-
-        service.Service.__init__(self, **params)
-
-    def command_line_args(self):
-        return ["--port=%d" % self.port] + self.service_args
+        service.ChromiumService.__init__(
+            self,
+            executable_path,
+            port,
+            service_args,
+            log_path,
+            env,
+            "Please download from https://developer.microsoft.com/en-us/microsoft-edge/tools/webdriver/")
