+++ py/selenium/webdriver/edge/webdriver.py
@@ -14,78 +14,49 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-import warnings
-
 from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
 from selenium.webdriver.edge.service import Service
-from selenium.webdriver.remote.remote_connection import RemoteConnection
-from selenium.webdriver.remote.webdriver import WebDriver as RemoteWebDriver
+from selenium.webdriver.chromium.webdriver import ChromiumDriver
+
 
 DEFAULT_PORT = 0
 DEFAULT_SERVICE_LOG_PATH = None
 
 
-class WebDriver(RemoteWebDriver):
+class WebDriver(ChromiumDriver):
 
     def __init__(self, executable_path='MicrosoftWebDriver.exe',
                  capabilities=None, port=DEFAULT_PORT, verbose=False,
                  service_log_path=None, log_path=DEFAULT_SERVICE_LOG_PATH,
-                 service=None, options=None, keep_alive=False):
+                 service=None, options=None, keep_alive=False, is_legacy=True,
+                 service_args=None):
         """
-        Creates a new instance of the chrome driver.
-
-        Starts the service and then creates new instance of chrome driver.
-
+        Creates a new instance of the edge driver.
+        Starts the service and then creates new instance of edge driver.
         :Args:
-         - executable_path - path to the executable. If the default is used it assumes the executable is in the $PATH
-         - capabilities - Dictionary object with non-browser specific
-           capabilities only, such as "proxy" or "loggingPref".
-         - port - port you would like the service to run, if left as 0, a free port will be found.
-         - verbose - whether to set verbose logging in the service
-         - service_log_path - Where to log information from the driver.
+         - executable_path - Deprecated: path to the executable. If the default is used it assumes the executable is in the $PATH
+         - capabilities - Dictionary object with non-browser specific capabilities only, such as "proxy" or "loggingPref".
+           Only available in Legacy mode
+         - port - Deprecated: port you would like the service to run, if left as 0, a free port will be found.
+         - verbose - whether to set verbose logging in the service. Only available in Legacy Mode
+         - service_log_path - Deprecated: Where to log information from the driver.
          - keep_alive - Whether to configure EdgeRemoteConnection to use HTTP keep-alive.
+         - service_args - Deprecated: List of args to pass to the driver service
+         - is_legacy: Whether to use MicrosoftWebDriver.exe (legacy) or MSEdgeDriver.exe (chromium-based). Defaults to True.
          """
-        if port != DEFAULT_PORT:
-            warnings.warn('port has been deprecated, please pass in a Service object',
-                          DeprecationWarning, stacklevel=2)
-        self.port = port
+        if not is_legacy:
+            executable_path = "msedgedriver"
 
-        if service_log_path != DEFAULT_SERVICE_LOG_PATH:
-            warnings.warn('service_log_path has been deprecated, please pass in a Service object',
-                          DeprecationWarning, stacklevel=2)
-        if capabilities is not None:
-            warnings.warn('capabilities has been deprecated, please pass in a Service object',
-                          DeprecationWarning, stacklevel=2)
-        if service_log_path != DEFAULT_SERVICE_LOG_PATH:
-            warnings.warn('service_log_path has been deprecated, please pass in a Service object',
-                          DeprecationWarning, stacklevel=2)
-        if verbose:
-            warnings.warn('verbose has been deprecated, please pass in a Service object',
-                          DeprecationWarning, stacklevel=2)
+        service = service or Service(executable_path,
+                                    port=port, verbose=verbose,
+                                    log_path=service_log_path, is_legacy=is_legacy)
 
-        if service:
-            self.service = service
-        else:
-            self.service = Service(executable_path, port=self.port, verbose=verbose,
-                                   log_path=service_log_path)
-        self.service.start()
-
-        if capabilities is None:
-            capabilities = DesiredCapabilities.EDGE
-
-        RemoteWebDriver.__init__(
-            self,
-            command_executor=RemoteConnection(self.service.service_url,
-                                              keep_alive=keep_alive),
-            desired_capabilities=capabilities)
-        self._is_remote = False
-
-    @property
-    def edge_service(self):
-        warnings.warn("'edge_service' has been deprecated, please use 'service'",
-                      DeprecationWarning, stacklevel=2)
-        return self.service
-
-    def quit(self):
-        RemoteWebDriver.quit(self)
-        self.service.stop()
+        super(WebDriver, self).__init__(
+            executable_path,
+            port,
+            options,
+            service_args,
+            DesiredCapabilities.EDGE,
+            service_log_path,
+            service,
+            keep_alive)
