+++ py/test/selenium/webdriver/common/driver_element_finding_tests.py
@@ -337,6 +337,7 @@ def test_Finding_ALink_By_Xpath_Using_Contains_Keyword_Should_Work(driver, pages
 
 
 @pytest.mark.xfail_chrome(raises=InvalidSelectorException)
+@pytest.mark.xfail_chromiumedge(raises=InvalidSelectorException)
 @pytest.mark.xfail_firefox(raises=InvalidSelectorException)
 @pytest.mark.xfail_remote(raises=InvalidSelectorException)
 @pytest.mark.xfail_marionette(raises=WebDriverException)
