+++ py/test/selenium/webdriver/common/frame_switching_tests.py
@@ -367,6 +367,7 @@ def testShouldBeAbleToSwitchToTheTopIfTheFrameIsDeletedFromUnderUsWithWebelement
 
 
 @pytest.mark.xfail_chrome(raises=NoSuchElementException)
+@pytest.mark.xfail_chromiumedge(raises=NoSuchElementException)
 @pytest.mark.xfail_marionette(raises=WebDriverException,
                               reason='https://github.com/mozilla/geckodriver/issues/614')
 @pytest.mark.xfail_remote(raises=WebDriverException,
