+++ py/test/selenium/webdriver/common/window_switching_tests.py
@@ -193,6 +193,7 @@ def testThatAccessingFindingAnElementAfterWindowIsClosedAndHaventswitchedDoesntC
 
 @pytest.mark.xfail_ie
 @pytest.mark.xfail_chrome
+@pytest.mark.xfail_chromiumedge
 def testShouldBeAbleToCreateANewWindow(driver, pages):
     original_handle = driver.current_window_handle
