+++ py/test/selenium/webdriver/common/window_tests.py
@@ -23,6 +23,7 @@ from selenium.webdriver.support.wait import WebDriverWait
 
 @pytest.mark.xfail_ie
 @pytest.mark.xfail_chrome(reason="Fails on Travis")
+@pytest.mark.xfail_chromiumedge(reason="Fails on Travis")
 @pytest.mark.xfail_marionette(reason="Fails on Travis")
 @pytest.mark.xfail_firefox(reason="Fails on Travis")
 @pytest.mark.xfail_remote(reason="Fails on Travis")
@@ -119,6 +120,8 @@ def test_should_set_the_rect_of_the_current_window(driver):
 
 @pytest.mark.xfail_chrome(raises=WebDriverException,
                           reason='Fullscreen command not implemented')
+@pytest.mark.xfail_chromiumedge(raises=WebDriverException,
+                          reason='Fullscreen command not implemented')
 @pytest.mark.xfail_firefox(raises=WebDriverException,
                            reason='Fullscreen command not implemented')
 @pytest.mark.xfail_safari(raises=WebDriverException,
@@ -144,6 +147,8 @@ def test_should_fullscreen_the_current_window(driver):
 
 
 @pytest.mark.xfail_chrome(raises=WebDriverException,
+                          reason='Minimize command not implemented')
+@pytest.mark.xfail_chromiumedge(raises=WebDriverException,
                           reason='Minimize command not implemented')
 @pytest.mark.xfail_firefox(raises=WebDriverException,
                            reason='Minimize command not implemented')
