+++ py/tox.ini
@@ -12,12 +12,13 @@ commands =
   py{27,36}-marionette: py.test --driver=Marionette {posargs}
   py{27,36}-remote: py.test --driver=Remote {posargs}
   py{27,36}-safari: py.test --driver=Safari {posargs}
+  py{27,36}-chromiumedge: py.test --driver=ChromiumEdge {posargs}
 install_command = pip install -v --no-index --find-links=../third_party/py {opts} {packages}
 deps =
   pytest==3.0.3
   pytest-instafail==0.3.0
   pytest-mock==1.5.0
-  py{27,33,34,35,36}-{unit,chrome,firefox,marionette}: pytest-xdist==1.15
+  py{27,33,34,35,36}-{unit,chrome,firefox,marionette,chromiumedge}: pytest-xdist==1.15
   urllib3==1.23
 
 [testenv:docs]
